Add stripPrefix middleware for path based routing
Some applications (notably Prometheus in this case) may not
correctly resolve endpoints if the prefix is passed through to
the application, as the "expected" endpoint per-unit is simply
`/` (for Prometheus `web.external-url`, including a prefix will
result in the Prometheus webroot itself also being at that URL).

Some flexibility will be required here, since other charms
(Grafana in particular) expect to see at least part of the prefix
if that is where they are hosted, or they will redirect users in
an infinite loop.

This is a first pass to unbreak Prometheus, and we can options from
there (whether it is dropping `web.external-url` from Prometheus,
keeping it as a constructor argument in Traefik, or other)