Fix has_ingress integration tests
Integration tests for `ingress-per-app` and `ingress-per-unit` were
validating that the ingress is present by calling `ping` with the `-p`
flag and the port. The `-p` flag is not for specifying a port however,
but to specify padding data to put in the payload of the `ICMP` packet.

Because the `metallb` configuration is using the local IP of the host
running `microk8s`, it is expected to always answer to pings. Therefore,
those tests were not testing anything.

This change replaces the `ping` with a direct socket, testing the TCP
connectivity to the ingress port.