Reduce the number of API calls to K8S
The charm makes a large number of API calls to retrieve information
about the IP address allocated to the K8S service; cache this
information to reduce the number of calls made and speed up hook
performance.

Note functools.lru_cache is used as this charm runs on 20.04 which
uses Python 3.8 - when the base is updated to 22.04 the more simple
cache decorator can be used instead.