fix: handle invalid combination of external_hostname and routing_path (#420)
Previously, this charm accepted the following configuration:
* routing_mode=subdomain
* external_hostname="" # (unset)

When external_hostname is unset, the url provided for any application related by ingress uses the LoadBalancer's external address, which may be an IP.  In these cases, it would provide charms urls like `model.app.1.2.3.4`, which are invalid URLs (the last segment of a URL cannot be all-numeric).  This led to an uncaught pydantic validation error when calling `ipa.publish_url()` because that method includes validation of the URL, putting the charm in Error state.  An example of this is shown in #361

Ideally, the fix here would be to validate the charm configuration+LoadBalancer details and halt charm execution if the configuration was invalid, putting the charm into BlockedStatus until resolved.  The problem is that the current architecture of this charm makes that solution challenging.  The charm is designed to atomically handle events (doing only the work a particular event needs) rather than holistically (recomputing the world on each event), meaning that skipping or losing track of events leads to undesired charm states.  Also, `ipa.publish_url()` is called deep in most (all?) event handlers, making it difficult to properly handle these errors at the charm level without a major refactor of the charm.

As a compromise, the following has been done:
* the traefik_k8s/v2/ingress.py library's `publish_url` method has been updated to catch the pydantic validation error cited in #361 and log it rather than raise it to the charm.  The library then writes ingress=None to the databag instead of the invalid URL, giving a soft indication to the user that the url is invalid.
* the config.yaml descriptions for routing_mode and external_hostname have been updated to explain the incompatibility in these settings
* config validation has been added to the __init__ of the charm for routing_mode and external_hostname.  If routing_mode==subdomain and hostname is unset, the charm will log warnings for the user about the possible incompatibility (but will not block the charm)

The upshot of these changes is that this charm will:
* not go into an unresponsive error state
* as best it can given the current charm architecture, warn the user of the misconfiguration
* not risk losing events through defer or getting event sequencing wrong

Fixes #361