Add support for healthcheck fields
This commit adds support for an optional dictionary
'healthcheck_params' containing healthcheck path, interval, and
timeout to be passed via an ingress relation. This allows Traefik
to detect down nodes and remove them from the loadbalancer rotation.
If 'healthcheck_params' is provided, the 'path' field is required from
the charm while 'interval' and 'timeout' can be omitted and will be
set to Traefik default values.