Add support for healthcheck (#460)
* Add support for healthcheck fields

This commit adds support for an optional dictionary
'healthcheck_params' containing healthcheck path, interval, and
timeout to be passed via an ingress relation. This allows Traefik
to detect down nodes and remove them from the loadbalancer rotation.
If 'healthcheck_params' is provided, the 'path' field is required from
the charm while 'interval' and 'timeout' can be omitted and will be
set to Traefik default values.

* Add support for healthcheck fields

This commit adds support for an optional dictionary
'healthcheck_params' containing healthcheck path, interval, and
timeout to be passed via an ingress relation. This allows Traefik
to detect down nodes and remove them from the loadbalancer rotation.
If 'healthcheck_params' is provided, the 'path' field is required from
the charm while 'interval' and 'timeout' can be omitted and will be
set to Traefik default values.

* switching to pydantic schema and adding itests

* remove duplicated params

* Remove duplicate keyword arguments

* fix attr error in utest

* copy libs into health tester charm

* Refactor IngressPerAppRequirer params docstring

* fix formatting

---------

Co-authored-by: IbraAoad <Ibrahim.Awwad@canonical.com>