feat: add support for consuming an ingress (#501)
Previously, this charm was only a provider of ingress.  The expected workflow was that users browse directly to this Traefik ingress URL and on through to the downstream applications.

This commit enables Traefik to be both a provider and *consumer* of the ingress relation, allowing us to layer more than one ingress.  For example, we can achieve:

```
App1 :ingress: Traefik1 :ingress: Traefik2 :ingress: ... :ingress: TraefikN [: User]
```

This can be useful for network topologies where some parts of the network are subdivided Traefik1 might ingress to a part of the subnet, while Traefik2 ingresses from external to Traefik1, etc.

Closes #396